refactor: replace prisma with supabase client
