+++ src/app/api/admin/users/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from 'next/server'
 import { getServerSession } from 'next-auth'
 import { authOptions } from '@/lib/auth'
-import { db } from '@/lib/db'
+import { db } from '@/lib/supabase'
 import { Prisma } from '@prisma/client'
 
 export async function GET(request: NextRequest) {
