+++ src/app/api/events/[id]/check-registration/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from 'next/server'
-import { db } from '@/lib/db'
+import { db } from '@/lib/supabase'
 
 export async function GET(
   request: NextRequest,
