+++ src/app/api/events/[id]/register/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from 'next/server'
-import { db } from '@/lib/db'
+import { db } from '@/lib/supabase'
 import { generateRegistrationToken } from '@/lib/tokens'
 
 export async function GET(
