+++ src/app/api/events/my-events/registrations/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from 'next/server'
 import { getServerSession } from 'next-auth'
 import { authOptions } from '@/lib/auth'
-import { db } from '@/lib/db'
+import { db } from '@/lib/supabase'
 
 export async function GET(_request: NextRequest) {
   try {
