+++ src/app/api/feedback/[feedbackId]/helpful/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from 'next/server'
-import { db } from '@/lib/db'
+import { db } from '@/lib/supabase'
 
 // POST /api/feedback/[feedbackId]/helpful - Ajouter ou supprimer un vote utile
 export async function POST(
