+++ src/app/api/organizer/events/[id]/delete/route.ts
@@ -1,7 +1,7 @@
 import { NextResponse } from 'next/server'
 import { getServerSession } from 'next-auth'
 import { authOptions } from '@/lib/auth'
-import { db } from '@/lib/db'
+import { db } from '@/lib/supabase'
 
 export async function DELETE(
   request: Request,
