+++ src/app/api/questions/[id]/votes/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from 'next/server'
-import { db } from '@/lib/db'
+import { db } from '@/lib/supabase'
 
 // POST /api/questions/[id]/votes - Ajouter ou modifier un vote
 export async function POST(
