+++ src/lib/auth.ts
@@ -1,6 +1,6 @@
 import NextAuth, { type NextAuthOptions } from 'next-auth'
 import CredentialsProvider from 'next-auth/providers/credentials'
-import { db } from '@/lib/db'
+import { db } from '@/lib/supabase'
 import bcrypt from 'bcryptjs'
 interface Logger {
   info(message: string, meta?: Record<string, unknown>): void
