+++ src/lib/db.ts
@@ -1,13 +1,4 @@
-import { PrismaClient } from '@prisma/client'
+import { db, supabase } from './supabase'
 
-const globalForPrisma = globalThis as unknown as {
-  prisma: PrismaClient | undefined
-}
-
-export const db =
-  globalForPrisma.prisma ??
-  (process.env.NODE_ENV === 'development'
-    ? new PrismaClient({ log: ['query'] })
-    : new PrismaClient());
-
-if (process.env.NODE_ENV !== 'production') globalForPrisma.prisma = db;
+export { db, supabase }
+export default db
