--- src/lib/prisma.ts
@@ -1,11 +1,0 @@
-import { PrismaClient } from '@prisma/client'
-
-declare global {
-  var prisma: PrismaClient | undefined
-}
-
-const prisma = globalThis.prisma || new PrismaClient()
-
-if (process.env.NODE_ENV !== 'production') globalThis.prisma = prisma
-
-export default prisma
