+++ src/lib/supabase.ts
@@ -1,0 +1,13 @@
+import { createClient } from '@supabase/supabase-js'
+
+const supabaseUrl = process.env.SUPABASE_URL ?? ''
+const serviceRoleKey = process.env.SUPABASE_SERVICE_ROLE_KEY ?? ''
+
+if (!supabaseUrl || !serviceRoleKey) {
+  throw new Error('Missing SUPABASE_URL or SUPABASE_SERVICE_ROLE_KEY')
+}
+
+export const supabase = createClient(supabaseUrl, serviceRoleKey)
+export const db = supabase
+
+export default supabase
